dj-backend-lq-main/pyproject.toml
[project]
name = "backend-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.30",
    "celery>=5.4.0",
    "celery-types>=0.23.0",
    "django>=5.2",
    "django-allauth[socialaccount]>=65.5.0",
    "django-allauth[headless]>=65.5.0",
    "django-cors-headers>=4.7.0",
    "django-currentuser>=0.9.0",
    "django-environ>=0.12.0",
    "django-json-widget>=2.0.1",
    "django-nested-admin>=4.1.0",
    "django-redis>=5.4.0",
    "django-storages>=1.14.6",
    "django-stubs[compatible-mypy]>=5.1.3",
    "djangorestframework>=3.15.0",
    "django-unfold==0.59.0",
    "django-zeal>=2.0.4",
    "graphene-django>=3.2.3",
    "graphene-django-cruddals==0.1.27",
    "gunicorn>=23.0.0",
    "ipython>=9.0.2",
    "mypy>=1.15.0",
    "pillow>=11.1.0",
    "psutil>=7.0.0",
    "psycopg>=3.2.6",
    "psycopg2-binary>=2.9.10",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "redis>=5.2.1",
    "requests>=2.32.0",
    "sentry-sdk[django]>=2.39.0",
    "shortuuid>=1.0.13",
    "static-ffmpeg>=2.13",
    "types-psutil>=7.0.0.20250401",
    "types-requests>=2.32.0.20250328",
    "uvicorn>=0.34.0",
    "djangorestframework-simplejwt>=5.5.1",
    "graphene-cruddals>=0.2.16",
    "openpyxl>=3.1.0",
    "whitenoise>=6.8.2",
    "pydantic>=2.12.5",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = ["ruff>=0.11.2", "flower>=2.0.1", "pre-commit>=4.2.0"]
